solved #4
