+++ client/src/components/Restaurants/Card.jsx
@@ -1,9 +1,10 @@
 import React, { useContext, useEffect, useState } from 'react'
 import axios from 'axios';
 import { Popup } from 'react-map-gl';
-import { AiFillStar, AiOutlineHeart, AiFillHeart } from 'react-icons/ai';
+import { AiFillStar, AiOutlineHeart } from 'react-icons/ai';
 import './Card.style.css'
 import { AppContext } from '../../App';
+import LikedButton from './LikedButton';
 
 const Card = ({data, opens, id, lat, long, setIsCurrent}) => {
 
@@ -16,18 +17,8 @@ const Card = ({data, opens, id, lat, long, setIsCurrent}) => {
     return open.id
   });
 
-  const [ favouritesList, setFavouritesList ] = useState([
-    {
-      email: '',
-      id: ''
-    }
-  ]);
-  const myFavourite = favouritesList.filter(list => list.email === currentUser.email);
-
-  const isFavourite = myFavourite.map(item => {
-    return item.id
-  }).includes(data.id)
-
+  const [ favouritesList, setFavouritesList ] = useState([]);
+  
   const addToFavourite = async(e) => {
     e.preventDefault();
     const categoriesArr = data.categories.map(category => {
@@ -45,15 +36,11 @@ const Card = ({data, opens, id, lat, long, setIsCurrent}) => {
       address,
       categories: categoriesArr
     }
-    if(loggedIn){
-      try {
-        const res = await axios.post('http://localhost:8888/api/favourites/add', newFavourite);
-        setFavouritesList([...favouritesList, res.data])
-      } catch (error) {
-        console.log(error)
-      }
-    }else{
-      alert('Please login.')
+    try {
+      const res = await axios.post('http://localhost:8888/api/favourites/add', newFavourite);
+      setFavouritesList([...favouritesList, res.data])
+    } catch (error) {
+      console.log(error)
     }
   }
 
@@ -119,15 +106,16 @@ const Card = ({data, opens, id, lat, long, setIsCurrent}) => {
           <div className='restaurant-address'>
             <p>{address}</p>
           </div>
-          {loggedIn && isFavourite ? (
-            <div className='unfavourite' onClick={removeFromFavourite}>
-              <AiFillHeart/>
-            </div>
+          {loggedIn && favouritesList.length > 0 ? (
+            <LikedButton 
+              favouritesList={favouritesList} 
+              addToFavourite={addToFavourite} 
+              removeFromFavourite={removeFromFavourite} 
+              data={data} 
+              currentUser={currentUser}/>
           ) : (
-            <div className='favourite' onClick={addToFavourite}>
-              <AiOutlineHeart/>
-            </div>
-            )}
+            <AiOutlineHeart onClick={() => alert('Please login.')}/>
+          )}
         </div>
       </div>
     </Popup>
